R3: extract private helpers, add Java interop rule to CLAUDE.md
- CLAUDE.md: document that extension functions on public API classes
  break Java callers' fluent chains; all public methods must be class members
- ReferenceLinkHelper.kt (new): deepCopyParameters, applyReferenceLinkRules,
  autoLinkReferences — extracted from linkReferences() method bodies
- FhirExtensionHelper.kt: add internal filterByExtension and
  filterByExtensionAndValueType top-level helpers
- OperationOutcomeExtensions.kt: add internal warningOutcome helper
- OperationResult.kt: delegate to extracted helpers, add class-level
  section map KDoc for navigation; remove IBaseHasExtensions import

OperationResult.kt reduced from 1526 → 1445 lines. All 454 tests pass.
